ENH add _utils.binning.bin_feature
